Deltaload and retmis bugfixes

+++ enquiries/script_RETMIS.py
@@ -29,45 +29,48 @@ def run_algo():
             sheet = workbook.active
 
             eb_sid = sheet["I2"].value
-            ec_sid = EnquiryComponentElements.objects.filter(eb_sid=eb_sid).first().ec_sid.ec_sid
-            task_enquiry_id = EnquiryComponentElements.objects.filter(eb_sid=eb_sid).first().ec_sid.erp_sid.cer_sid.enquiry_id
+            if EnquiryComponentElements.objects.filter(eb_sid=eb_sid).exists():
+                ec_sid = EnquiryComponentElements.objects.filter(eb_sid=eb_sid).first().ec_sid.ec_sid
+                task_enquiry_id = EnquiryComponentElements.objects.filter(eb_sid=eb_sid).first().ec_sid.erp_sid.cer_sid.enquiry_id
 
-            #TODO add safety checks on file content (or lock down file)
-            task_pk = None
+                #TODO add safety checks on file content (or lock down file)
+                task_pk = None
 
-            try:
-                task_pk = TaskManager.objects.get(task_id='RETMIS', ec_sid=ec_sid ,task_completion_date__isnull=True).pk
-            except:
-                pass
-            if task_pk is not None:
-                MisReturnData.objects.create(
-                    eb_sid = EnquiryBatches.objects.get(eb_sid=eb_sid),
-                    ec_sid = EnquiryComponents.objects.get(ec_sid=ec_sid),
-                    original_exm = sheet["D4"].value,
-                    rev_exm = sheet["E4"].value,
-                    original_mark = sheet["F4"].value,
-                    mark_status = sheet["G4"].value,
-                    revised_mark = sheet["H4"].value,
-                    justification_code = sheet["I4"].value,
-                    remark_reason = sheet["B40"].value,
-                    remark_concern_reason = sheet["B50"].value
-                )
+                try:
+                    task_pk = TaskManager.objects.get(task_id='RETMIS', ec_sid=ec_sid ,task_completion_date__isnull=True).pk
+                except:
+                    pass
+                if task_pk is not None:
+                    MisReturnData.objects.create(
+                        eb_sid = EnquiryBatches.objects.get(eb_sid=eb_sid),
+                        ec_sid = EnquiryComponents.objects.get(ec_sid=ec_sid),
+                        original_exm = sheet["D4"].value,
+                        rev_exm = sheet["E4"].value,
+                        original_mark = sheet["F4"].value,
+                        mark_status = sheet["G4"].value,
+                        revised_mark = sheet["H4"].value,
+                        justification_code = sheet["I4"].value,
+                        remark_reason = sheet["B40"].value,
+                        remark_concern_reason = sheet["B50"].value
+                    )
 
-                #Move file to completed folder
-                shutil.move(filename, new_filename)
+                    #Move file to completed folder
+                    shutil.move(filename, new_filename)
 
-                #Create next step in chain (MISVRM)
-                TaskManager.objects.create(
-                    enquiry_id = CentreEnquiryRequests.objects.get(enquiry_id=task_enquiry_id),
-                    ec_sid = EnquiryComponents.objects.get(ec_sid=ec_sid),
-                    task_id = 'MISVRM',
-                    task_assigned_to = None,
-                    task_assigned_date = None,
-                    task_completion_date = None
-                )
-                #complete the task
-                TaskManager.objects.filter(pk=task_pk,task_id='RETMIS').update(task_completion_date=timezone.now())
-                ScriptApportionment.objects.filter(ec_sid=ec_sid).update(script_marked=0)
+                    #Create next step in chain (MISVRM)
+                    TaskManager.objects.create(
+                        enquiry_id = CentreEnquiryRequests.objects.get(enquiry_id=task_enquiry_id),
+                        ec_sid = EnquiryComponents.objects.get(ec_sid=ec_sid),
+                        task_id = 'MISVRM',
+                        task_assigned_to = None,
+                        task_assigned_date = None,
+                        task_completion_date = None
+                    )
+                    #complete the task
+                    TaskManager.objects.filter(pk=task_pk,task_id='RETMIS').update(task_completion_date=timezone.now())
+                    ScriptApportionment.objects.filter(ec_sid=ec_sid).update(script_marked=0)
+            else:
+                print('FAILED:' + eb_sid)
         else:
             if file.endswith("COMPLETE") or file.endswith("FILE_CHECKS"):
                 a=1 #do nothing
